refactor: move rejection tests to separate class
